split common.js into two files, test new import on test files
