+++ src/server/test/commonTestDB.js
@@ -1,0 +1,91 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+/* This file exports two useful items: testDB, which provides the method .getConnection,
+ * returning a connection to the testing database, and recreateDB, which destroys the
+ * database and creates a new schema there.
+ */
+
+const { log, LogLevel } = require('../log');
+
+// TODO: Move logging controls to a better place.
+// During testing do not show errors since many tests generate them.
+log.level = LogLevel.SILENT;
+// If tests are failing then turning on the console messages can help. Commenting the line
+// above and uncommenting the next line will do that.
+// log.level = LogLevel.DEBUG;
+// During testing do not do emails. Since DB wiped and not set up for emails this would not do anything.
+log.emailLevel = LogLevel.SILENT;
+// Logging to the DB during testing has issues. First, the test DB is wiped before each test so it does not actually
+// stay around. Second, if one test finishes
+// quickly after the DB logging starts (async) then the next test starts and the beforeEach wipes the DB.
+// If the attempt to put the log msg in the DB
+// happens before it is initialized again then the log table is missing and causes an error.
+log.logToDb = false;
+
+const User = require('../models/User');
+const { getDB, createSchema, stopDB } = require('../models/database');
+const { swapConnection, dropConnection } = require('../db');
+const app = require('../app');
+const bcrypt = require('bcryptjs');
+const { mocha, chai, expect } = require('./common');
+
+// Global database connection object, whose members are accessible to DB-touching tests.
+// For instance, call:
+// conn = testDB.getConnection();
+// doSomething(conn)
+let testDB = {
+	_connection: null,
+	config: null,
+	getConnection: function () {
+		return this._connection;
+	}
+};
+
+function connectTestDB() {
+	const testDBConfig = {
+		user: process.env.OED_DB_TEST_USER || process.env.OED_DB_USER,
+		database: process.env.OED_DB_TEST_DATABASE,
+		password: process.env.OED_DB_TEST_PASSWORD || process.env.OED_DB_PASSWORD,
+		host: process.env.OED_DB_TEST_HOST || process.env.OED_DB_HOST,
+		port: process.env.OED_DB_TEST_PORT || process.env.OED_DB_PORT
+	};
+
+	stopDB();
+	testDB._connection = getDB(testDBConfig);
+	testDB.config = testDBConfig;
+	swapConnection(testDB.config, testDB._connection);
+}
+
+// The user for use by tests.
+const testUser = new User(undefined, 'test@example.invalid', bcrypt.hashSync('password', 10), User.role.ADMIN);
+testUser.password = 'password';
+
+async function recreateDB() {
+	const conn = testDB.getConnection();
+	// This should drop all database objects, as long as they were all created by the current database user
+	// They should be, since they were all created during a previous test.
+	await conn.none('DROP OWNED BY current_user;');
+	await createSchema(conn);
+	await testUser.insert(conn);
+}
+
+mocha.before(() => {
+	connectTestDB();
+});
+
+mocha.beforeEach(async () => {
+	await recreateDB();
+});
+
+mocha.after(() => {
+	dropConnection();
+});
+
+module.exports = {
+	app,
+	testUser,
+	recreateDB,
+	testDB
+};
