+++ src/server/test/routes/baselineParamsTest.js
@@ -4,7 +4,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-const { chai, mocha, app, expect } = require('../common');
+//const { chai, mocha, app, expect } = require('../common');
+const { chai, mocha, expect } = require('../common');
+const { app } = require('../commonTestDB');
 const { HTTP_CODES } = require('../../util/httpCodes');
 
 mocha.describe('Baseline Parameter Validation', () => {
