+++ src/server/test/routes/conversionsParamsTest.js
@@ -4,7 +4,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-const { chai, mocha, app, expect } = require('../common');
+//const { chai, mocha, app, expect } = require('../common');
+const { chai, mocha, expect } = require('../common');
+const { app } = require('../commonTestDB');
 const { testInvalidField, validateInt, validateBool, validateNoExtraFields } = require('../util/validationHelpers');
 const { STRING_GENERAL_MAX_LENGTH } = require('../../util/validationConstants');
 const { HTTP_CODES } = require('../../util/httpCodes');
